requirements.txt
dj-database-url==0.5.0
Django==2.1.8
django-cors-headers==3.0.2
djangorestframework==3.9.4
gunicorn==19.9.0
psycopg2==2.8.3
Pygments==2.4.2
pytz==2019.1
whitenoise==4.1.2
